🚧 First db proper calls + files structure + table

+++ php/initializeProject.php
@@ -4,11 +4,14 @@ if (isset($_GET['demo_data'])) {
 } else {
     $demo = false;
 }
-$db = new SQLite3('db/fsm.db');
-// Tables
-// -- Users table
-$db->exec("DROP TABLE IF EXISTS users");
-$db->exec("CREATE TABLE users(
+if (!isset($_GET['iKnowWhatIamDoing'])) {
+    echo "Please read the docs and do not call this file unless you know what you are doing!";
+} else {
+    $db = new SQLite3('../db/fsm.db');
+    // Tables
+    // -- Users table
+    $db->exec("DROP TABLE IF EXISTS users");
+    $db->exec("CREATE TABLE users(
 id INTEGER PRIMARY KEY, 
 name TEXT, 
 email TEXT, 
@@ -21,7 +24,7 @@ admin INTEGER DEFAULT 0,
 active INTEGER DEFAULT 1)");
 // ---- Test Data
 if ($demo) {
-$db->exec("INSERT INTO users(name, email, phone, uuid, pw, email_recap, email_recap_frequency, admin) 
+    $db->exec("INSERT INTO users(name, email, phone, uuid, pw, email_recap, email_recap_frequency, admin) 
 VALUES('Developer A', 'cntac@dfs.com', '00312545542', '123b04f6-0243-4ed8-bdeb-2095c14c578e', 'fafwae', 1, 2, 1)");
 $db->exec("INSERT INTO users(name, email, phone, uuid, pw, email_recap, email_recap_frequency, admin) 
 VALUES('Developer B', 'cntsac@dfda.fr', '0031526142', '92d2ff16-8642-400a-a905-9d70e656274c', 'favcbngae', 0, 0, 1)");
@@ -45,7 +48,7 @@ auto_create_month INT DEFAULT 0, -- Month on which yearly payments are created (
 active INT DEFAULT 1)");
 // ---- Test Data
 if ($demo) {
-$db->exec("INSERT INTO subscriptions(name, actual_price, actual_payers, payment_frequency, auto_create, auto_create_day) 
+    $db->exec("INSERT INTO subscriptions(name, actual_price, actual_payers, payment_frequency, auto_create, auto_create_day) 
 VALUES('Netflix', 15.99, 5, 3, 1, 1)");
 $db->exec("INSERT INTO subscriptions(name, actual_price, actual_payers, payment_frequency, auto_create, auto_create_day) 
 VALUES('Spotify', 17.99, 6, 3, 1, 1)");
@@ -53,7 +56,7 @@ $db->exec("INSERT INTO subscriptions(name, actual_price, actual_payers, payment_
 VALUES('Disney Plus', 74.5, 5, 4, 1, 15, 9)");
 }
 
-// -- Payments table
+// -- Payments to service table
 $db->exec("DROP TABLE IF EXISTS subscriptionPayments");
 $db->exec("CREATE TABLE subscriptionPayments(
 id INTEGER PRIMARY KEY, 
@@ -64,7 +67,7 @@ comments TEXT,
 FOREIGN KEY(subscription) REFERENCES subscriptions(id))");
 // ---- Test Data
 if ($demo) {
-$db->exec("INSERT INTO subscriptionPayments(subscription, date, payers, comments) 
+    $db->exec("INSERT INTO subscriptionPayments(subscription, date, payers, comments) 
 VALUES(1, 1680310800, 3, 'all good')");
 $db->exec("INSERT INTO subscriptionPayments(subscription, date, payers, comments) 
 VALUES(1, 1682902800, 3, 'not so great')");
@@ -72,7 +75,38 @@ $db->exec("INSERT INTO subscriptionPayments(subscription, date, payers, comments
 VALUES(1, 1685581200, 3, 'ok')");
 }
 
-// Make check if data is available ? DO NOTHING unless you have a certain php parameter : Initialize DB 
+// -- Payments of co-subscribers
+$db->exec("DROP TABLE IF EXISTS cosubscribersPayments");
+$db->exec("CREATE TABLE cosubscribersPayments(
+id INTEGER PRIMARY KEY, 
+subscriptionPayment INTEGER, -- Foreign id to subscriptions.id
+payer INTEGER, -- Foreign id to users.id
+date INTEGER, 
+payed INT DEFAULT 0, 
+comments TEXT,
+FOREIGN KEY(subscriptionPayment) REFERENCES subscriptionPayments(id),
+FOREIGN KEY(payer) REFERENCES users(id))");
+// ---- Test Data
+if ($demo) {
+$db->exec("INSERT INTO cosubscribersPayments(subscriptionPayment, payer, date, payed, comments) 
+VALUES(1, 2, 1680310800, 1, 'yes')");
+$db->exec("INSERT INTO cosubscribersPayments(subscriptionPayment, payer, date, payed, comments) 
+VALUES(1, 3, 1680310800, 1, 'maybe')");
+$db->exec("INSERT INTO cosubscribersPayments(subscriptionPayment, payer, date, payed, comments) 
+VALUES(1, 4, 1680310800, 1, 'dragon')");
+$db->exec("INSERT INTO cosubscribersPayments(subscriptionPayment, payer, date, payed, comments) 
+VALUES(2, 2, 1682902800, 1, 'ground')");
+$db->exec("INSERT INTO cosubscribersPayments(subscriptionPayment, payer, date, payed, comments) 
+VALUES(2, 3, 1682902800, 0, 'pirate')");
+$db->exec("INSERT INTO cosubscribersPayments(subscriptionPayment, payer, date, payed, comments) 
+VALUES(3, 2, 1685581200, 0, 'team')");
+$db->exec("INSERT INTO cosubscribersPayments(subscriptionPayment, payer, date, payed, comments) 
+VALUES(3, 3, 1685581200, 0, 'sheldon')");
+}
+
+echo "The DB was created successfully. Please go back to the <a href='../'>app</a> now.";
+}
+
 
 // Send email welcoming to project
